requirements.txt
Flask
matplotlib
PyYAML
requests
httpx[cli,http2]
openai